+++ src/Models/OutgoingRequestTrace.php
@@ -2,6 +2,7 @@
 
 namespace KolayBi\RequestTracer\Models;
 
+use Illuminate\Database\Eloquent\Concerns\HasUlids;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Carbon;
 
@@ -10,6 +11,8 @@ use Illuminate\Support\Carbon;
  */
 class OutgoingRequestTrace extends Model
 {
+    use HasUlids;
+
     public $timestamps = false;
 
     protected $guarded = [];